Import Air Quality Site Metadata from Dataflow D

Imports data from the data flow D data viewer. This is useful for binding onto datasets obtained using get_eea_parquet_files() to assign monitoring data with relevant site information. Metadata can be filtered by providing the countries and cities arguments.

Usage

import_eea_stations(countries = NULL, cities = NA, ..., .cache = TRUE)

Arguments

countries

A vector of country codes from import_eea_countries(). If NULL, data from all countries will be imported.

cities

A vector of cities in the given countries from import_eea_cities(). If NULL, data from all cities in the given countries will be imported.

...

Not currently used; reserved for future functionality.

.cache

This function works by installing a CSV to a temporary directory and reading it into R. When .cache is TRUE, this only occurs once per R session. When .cache is FALSE, this CSV is re-downloaded every time the function is called.

Author

Jack Davison